+++ src/components/MapViewer/MapViewer.tsx
@@ -9,7 +9,7 @@ import BottomDrawer from '../Drawer/BottomDrawer'
 import RightDrawer from '../Drawer/RightDrawer'
 import { DrawerBase } from '../Drawer/DrawerContext'
 import { MapViewerBase, useMapViewerContext } from './MapViewerContext'
-import LayerTree from './LayerTree'
+import LayerTree from '../LayerTree/LayerTree'
 import { TabListBase } from '../Tab/TabListContext'
 import { Tabs } from '../Tab/Tabs'
 import ListGuesser from '../../jsonapi/components/ListGuesser'
